changing methods to return single write connection
it makes sense for each connection group to have a single write connection.
by guaranteeing this, we can clean up a lot of the code by having it grab the
single write connection and perform an action. this removes a lot of looping
logic, making it easier to understand what is happening.